refactor(remote-config): migrate rcstatus component to v2
