+++ comp/remote-config/bundle.go
@@ -8,7 +8,7 @@ package remoteconfig
 
 import (
 	"github.com/DataDog/datadog-agent/comp/remote-config/rcclient/rcclientimpl"
-	"github.com/DataDog/datadog-agent/comp/remote-config/rcstatus/rcstatusimpl"
+	rcstatusfx "github.com/DataDog/datadog-agent/comp/remote-config/rcstatus/fx"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
@@ -18,6 +18,6 @@ import (
 func Bundle() fxutil.BundleOptions {
 	return fxutil.Bundle(
 		rcclientimpl.Module(),
-		rcstatusimpl.Module(),
+		rcstatusfx.Module(),
 	)
 }
