+++ comp/remote-config/rcstatus/def/component.go
