+++ comp/remote-config/rcstatus/impl/status_templates/remoteconfiguration.tmpl
